Refine Agent popup and local CI gates
