+++ hushh-webapp/__tests__/setup.ts
@@ -11,6 +11,8 @@ import { vi, beforeEach } from "vitest";
 // Mock environment variables for testing
 // The 'process' object is now recognized thanks to the node reference above
 process.env.NEXT_PUBLIC_APP_ENV = "development";
+process.env.NEXT_PUBLIC_FIREBASE_API_KEY =
+  "AIzaSyDummylocaltestkey000000000000000000";
 process.env.BACKEND_URL = "http://localhost:8000";
 process.env.NODE_ENV = "test";
