+++ hushh-webapp/lib/services/one-kyc-client-zk-service.ts
@@ -285,6 +285,7 @@ function formatDateTimeValue(value: string): string | null {
   return new Intl.DateTimeFormat("en-US", {
     dateStyle: "medium",
     timeStyle: "short",
+    timeZone: "UTC",
   }).format(date);
 }
